+++ changelog.md
@@ -1,5 +1,10 @@
 # Version history of thermos
 
+## Version 0.2.0 (2022-04-24)
+
+`thermo-logger`, a command line application that continuously logs thermal
+sensor data to a file, is introduced.
+
 ## Version 0.1.1 (2022-04-24)
 
 The file name matching algorithm for reading thermal devices from